rag-with-citations/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
        "arxiv>=2.1.3",
    "faiss-cpu>=1.10.0",
    "google-genai>=1.7.0",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.9",
    "langsmith>=0.3.17",
    "llama-index>=0.12.25",
    "pypdf>=5.4.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]
